Refactor block test specs
